+++ src/app/services/api.js
@@ -57,6 +57,9 @@
         };
 
 
+        this.getHost = () => {
+            return  (connection.host);
+        };
         this.getPassword = () => {
             return  (connection.password);
         };
@@ -166,6 +169,9 @@
                 if (!(connection.host.indexOf('://') > 0 || connection.host.indexOf('/') == 0)) {
                     httpProto = 'http://';
                 }
+                // ClickHouse/dbms/src/Interpreters/Settings.h : https://github.com/yandex/ClickHouse/blob/master/dbms/src/Interpreters/Settings.h
+
+
 
                 url = httpProto + connection.host + '/?add_http_cors_header=1&log_queries=1&output_format_json_quote_64bit_integers=0&output_format_json_quote_denormals=1';
                 //max_block_size=1&send_progress_in_http_headers=1&http_headers_progress_interval_ms=500